sqlite: add fake update command (#7)
Adds fake `update` command which is a no-op as SQLite database is
supposed to be readonly.
The main reason is to be able to demonstrate Pomerium blocking
individual tool calls.

i.e. given the route defined as

```yaml
routes:
  - from: https://db.mcp.sandbox.pomerium.io
    to: http://mcp-servers:8080/sqlite
    name: Northwind DB
    mcp:
      server: {}
    policy:
      - allow:
          and:
            - domain:
                is: pomerium.com
            - mcp_tool:
                in: ["describe_table", "list_tables", "read_query"]
```

If you try to ask an LLM to update the table, this request would be
blocked by Pomerium.